fix: profile 에 관련된 api로직 server로 이동완료
