+++ apps/client/.turbo/turbo-build.log
@@ -2,19 +2,7 @@ info  - Loaded env from /Users/chanhyuk/Desktop/toppings-client/apps/client/.env
 info  - Linting and checking validity of types...
 error - ESLint must be installed in order to run during builds: yarn add --dev eslint
 info  - Creating an optimized production build...
-warn  - Compiled with warnings
-
-../../.yarn/cache/axios-npm-1.2.0-ad83373c57-f08ce214e9.zip/node_modules/axios/lib/helpers/toFormData.js
-A Node.js API is used (Buffer at line: 138) which is not supported in the Edge Runtime.
-Learn more: https://nextjs.org/docs/api-reference/edge-runtime
-
-Import trace for requested module:
-../../.yarn/cache/axios-npm-1.2.0-ad83373c57-f08ce214e9.zip/node_modules/axios/lib/helpers/toFormData.js
-../../.yarn/cache/axios-npm-1.2.0-ad83373c57-f08ce214e9.zip/node_modules/axios/lib/axios.js
-../../.yarn/cache/axios-npm-1.2.0-ad83373c57-f08ce214e9.zip/node_modules/axios/index.js
-./constants/Api.ts
-./constants/index.ts
-
+info  - Compiled successfully
 info  - Collecting page data...
 info  - Generating static pages (0/19)
 info  - Generating static pages (4/19)
@@ -31,27 +19,27 @@ Route (pages)                              Size     First Load JS
 ├ ○ /login/redirect                        272 B           357 kB
 ├ ○ /map                                   3.98 kB         361 kB
 ├ ○ /notice                                276 B           357 kB
-├ λ /post/[id]                             954 B           364 kB
-├ ○ /post/add                              971 B           364 kB
-├ λ /profile                               2.24 kB         359 kB
-├ λ /profile/edit                          3.74 kB         360 kB
+├ λ /post/[id]                             956 B           364 kB
+├ ○ /post/add                              973 B           364 kB
+├ λ /profile                               3.04 kB         360 kB
+├ λ /profile/edit                          3.57 kB         360 kB
 ├ ○ /profile/edit/eatingHabits             2.38 kB         359 kB
-├ ○ /profile/edit/nationality              2.4 kB          359 kB
-├ ○ /profile/menu                          2.15 kB         359 kB
-├ λ /profile/posts                         1.78 kB         359 kB
+├ ○ /profile/edit/nationality              2.39 kB         359 kB
+├ ○ /profile/menu                          2.08 kB         359 kB
+├ λ /profile/posts                         2.56 kB         359 kB
 ├ ○ /profile/reviews                       265 B           357 kB
 ├ ○ /profile/saved                         263 B           357 kB
 ├ ○ /recent                                2.12 kB         362 kB
 ├ ○ /recent/eatingHabit                    2.02 kB         362 kB
-├ ○ /recent/nationality                    2.02 kB         362 kB
+├ ○ /recent/nationality                    2.03 kB         362 kB
 ├ ○ /recent/restaurant                     1.22 kB         361 kB
-├ ○ /register/eatingHabits                 4 kB            361 kB
+├ ○ /register/eatingHabits                 3.87 kB         361 kB
 ├ ○ /register/nationality                  2.35 kB         359 kB
-└ λ /search/[type]                         1.83 kB         359 kB
+└ λ /search/[type]                         1.84 kB         359 kB
 + First Load JS shared by all              357 kB
   ├ chunks/framework-749f74b9b46c2a3f.js   45.4 kB
   ├ chunks/main-e2980877f7152e6b.js        30.7 kB
-  ├ chunks/pages/_app-aa3a03db559af102.js  279 kB
+  ├ chunks/pages/_app-29f8c54301cba848.js  280 kB
   └ chunks/webpack-235955b20eb3ee85.js     1.15 kB
 
 ƒ Middleware                               62 kB
