+++ apps/client/apis/index.ts
@@ -1,2 +1,1 @@
 export * from "./register";
-export * from "./profile";
