+++ apps/client/components/Profile/edit/UserInfo/ProfileImage/index.tsx
@@ -6,7 +6,7 @@ import { OrangePlus } from "@svgs/common";
 import { flex, gutter, touchable } from "@toss/emotion-utils";
 import Image from "next/image";
 import { useId } from "react";
-import { useFetchUserInfo } from "~/queries/profile";
+import { useFetchUserInfo } from "~/server/profile";
 import { imageUploader } from "~/utils";
 
 const ProfileImage = () => {
