+++ apps/client/components/Profile/posts/UserBadge/index.tsx
@@ -3,7 +3,7 @@ import { avatar } from "@images/profile";
 import { Flex, gutter, margin, padding, size } from "@toss/emotion-utils";
 import Image from "next/image";
 import { Text } from "~/components/Common/Typo";
-import { useFetchUserInfo } from "~/queries/profile";
+import { useFetchUserInfo } from "~/server/profile";
 import { hexToRgba } from "~/utils";
 
 const UserBadge = () => {
