--- apps/client/mutations/profile/index.ts
@@ -1,2 +1,0 @@
-export { default as useUpdateUserInfo } from "./useUpdateUserInfo";
-export { default as useLogout } from "./useLogout";
