--- apps/client/mutations/profile/useLogout.tsx
@@ -1,9 +1,0 @@
-import type { UseQueryOptions } from "@tanstack/react-query";
-import { useMutation } from "@tanstack/react-query";
-import { logout } from "~/apis";
-
-const useLogout = (options: Pick<UseQueryOptions, "onSuccess">) => {
-  return useMutation(logout, options);
-};
-
-export default useLogout;
