--- apps/client/mutations/profile/useUpdateUserInfo.ts
@@ -1,8 +1,0 @@
-import type { UseQueryOptions } from "@tanstack/react-query";
-import { useMutation } from "@tanstack/react-query";
-import { updateUserInfo } from "~/apis";
-
-const useUpdateUserInfo = (options: Pick<UseQueryOptions, "onSuccess">) => {
-  return useMutation(updateUserInfo, options);
-};
-export default useUpdateUserInfo;
