+++ apps/client/pages/profile/edit/index.page.tsx
@@ -12,8 +12,7 @@ import { Text } from "~/components/Common/Typo";
 import { UserInfo } from "~/components/Profile/edit";
 import { env } from "~/constants";
 import { useInternalRouter, useSetNavigation } from "~/hooks";
-import { useUpdateUserInfo } from "~/mutations/profile";
-import { Keys, useFetchUserInfo } from "~/queries/profile";
+import { Keys, useFetchUserInfo, useUpdateUserInfo } from "~/server/profile";
 
 const ProfileEdit = () => {
   const { push } = useInternalRouter();
@@ -35,7 +34,7 @@ const ProfileEdit = () => {
   const { data } = useFetchUserInfo();
   const edit = useEditValue();
 
-  const { mutate } = useUpdateUserInfo({
+  const { mutate: updateUserInfoMutate } = useUpdateUserInfo({
     onSuccess: () => {
       overlay.open(() => <SuccessModal />);
 
@@ -46,7 +45,7 @@ const ProfileEdit = () => {
   });
 
   const onClickRegisterHandler = useCallback(() => {
-    mutate({
+    updateUserInfoMutate({
       name: edit.name ?? data.name,
       country: edit.country ?? data.country,
       habits: edit.habits ?? data.habits,
@@ -61,7 +60,7 @@ const ProfileEdit = () => {
     edit.habits,
     edit.name,
     edit.profile,
-    mutate
+    updateUserInfoMutate
   ]);
 
   return (
