--- apps/client/queries/profile/index.ts
@@ -1,3 +1,0 @@
-export { default as Keys } from "./keys";
-export { default as useFetchUserInfo } from "./useFetchUserInfo";
-export { default as useFetchUserPosts } from "./useFetchUserPosts";
