--- apps/client/queries/profile/keys.ts
@@ -1,6 +1,0 @@
-const Keys = {
-  user: () => ["user"],
-  posts: () => ["posts"]
-};
-
-export default Keys;
