+++ apps/client/server/profile/keys.ts
@@ -1,0 +1,6 @@
+const Keys = {
+  user: () => ["user"],
+  posts: () => ["posts"]
+};
+
+export default Keys;
