+++ apps/client/server/profile/mutations.ts
@@ -1,0 +1,13 @@
+import type { UseQueryOptions } from "@tanstack/react-query";
+import { useMutation } from "@tanstack/react-query";
+import { logout, updateUserInfo } from "./apis";
+
+export const useLogout = (options: Pick<UseQueryOptions, "onSuccess">) => {
+  return useMutation(logout, options);
+};
+
+export const useUpdateUserInfo = (
+  options: Pick<UseQueryOptions, "onSuccess">
+) => {
+  return useMutation(updateUserInfo, options);
+};
